mi commit!

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const mongodb = require('mongodb');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
 
@@ -68,9 +69,24 @@ exports.getAddProduct = (req, res, next) => {
 // };
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        });
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log(errors.array());
@@ -89,6 +105,7 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+    const imageUrl = image.path;
     const product = new Product({
         // _id: mongoose.Types.ObjectId('5bfd02a57ebe953f72d59686'),
         title: title,
@@ -309,7 +326,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -320,7 +337,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -336,7 +352,10 @@ exports.postEditProduct = (req, res, next) => {
         }
         product.title = updatedTitle;
         product.price = updatedPrice;
-        product.imageUrl = updatedImageUrl;
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         product.description = updatedDescription;
         return product.save()
         .then(result => {
@@ -382,9 +401,16 @@ exports.postEditProduct = (req, res, next) => {
 // };
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({
-        _id: prodId,
-        userId: req.user._id
+    Product.findById(prodId)
+    .then(product => {
+        if (!product) {
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({
+            _id: prodId,
+            userId: req.user._id
+        });
     })
     .then(() => {
         console.log('Destroyed product');
